refact : add folder and move jwt related classes
