+++ travelfeeldog/src/main/java/travelfeeldog/member/domain/application/service/MemberReadWriteService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
-import travelfeeldog.global.auth.jwt.JwtProvider;
-import travelfeeldog.global.auth.jwt.TokenResponse;
+import travelfeeldog.global.auth.jwt.service.JwtProvider;
+import travelfeeldog.global.auth.jwt.response.TokenResponse;
 import travelfeeldog.member.dto.MemberDto;
 import travelfeeldog.member.dto.MemberDtos.*;
 import travelfeeldog.member.dto.MemberNickNameHistoryDto;
