+++ travelfeeldog/src/main/java/travelfeeldog/member/domain/application/service/MemberWriteService.java
@@ -1,6 +1,6 @@
 package travelfeeldog.member.domain.application.service;
 
-import travelfeeldog.global.auth.jwt.TokenResponse;
+import travelfeeldog.global.auth.jwt.response.TokenResponse;
 import travelfeeldog.member.domain.model.Member;
 import travelfeeldog.member.dto.MemberDtos.MemberPostRequestDto;
 import travelfeeldog.member.dto.MemberDtos.MemberPostResponseDto;
