+++ travelfeeldog/src/main/java/travelfeeldog/member/domain/model/Member.java
@@ -26,7 +26,6 @@ import org.springframework.util.Assert;
 import travelfeeldog.community.feedlike.domain.model.FeedLike;
 import travelfeeldog.community.feed.domain.model.Feed;
 import travelfeeldog.community.scrap.domain.model.Scrap;
-import travelfeeldog.global.auth.jwt.TokenResponse;
 import travelfeeldog.review.domain.review.model.Review;
 import travelfeeldog.global.common.domain.basetime.BaseTimeEntity;
